changes runtests to run python server tests
