+++ test/client_errors.jl
@@ -15,125 +15,102 @@ host = hostname(server)
 upload_file_name = "test_upload.txt"
 file_name = "test_download.txt"
 
-# @testset "FTPClientError" begin
 
-#     @testset "ftp() do ftp_client end" begin
-#         ftp(ssl=false, user=user, pswd=pswd, host=host) do f
-#             @test_throws FTPClientError download(f, file_name)
-#        end
+# ftp() do ftp_client end
+#ftp(ssl=false, user=user, pswd=pswd, host=host) do f
+#@test_throws FTPClientError download(f, file_name)
+#end
 
-#     end
 
-#       ftp_init()
+ftp_init()
 
-#     @testset "ftp_connect error" begin
+# make error
 
-         buff = IOBuffer()
-         msg = "This will go into the message"
-         lib_curl_error = 765
-         error = FTPClientError(msg, lib_curl_error)
-         showerror(buff, error)
-         seekstart(buff)
-         @test "$msg :: LibCURL error #$lib_curl_error" == readstring(buff)
+buff = IOBuffer()
+msg = "This will go into the message"
+lib_curl_error = 765
+error = FTPClientError(msg, lib_curl_error)
+showerror(buff, error)
+seekstart(buff)
+@test "$msg :: LibCURL error #$lib_curl_error" == readstring(buff)
 
-#     end
 
-#     @testset "Testing for client failure in FTPC.jl" begin
 
-#         @testset "ftp_connect error" begin
-             options = RequestOptions(ssl=false, active_mode=false, username=user, passwd=pswd, hostname="not a host")
-             @test_throws FTPClientError ftp_connect(options)
-#         end
+# connect error
+options = RequestOptions(ssl=false, active_mode=false, username=user, passwd=pswd, hostname="not a host")
+@test_throws FTPClientError ftp_connect(options)
 
-#         @testset "ftp_put error" begin
-#             options = RequestOptions(ssl=false, active_mode=false, username=user, passwd=pswd, hostname=host)
-#             open(upload_file_name) do file
-#                 @test_throws FTPClientError ftp_put(upload_file_name, file, options)
-#             end
-#         end
+# FTP object error when connecting
+@test_throws FTPClientError FTP(ssl=false, user=user, pswd=pswd, host="not a host")
 
-#         @testset "ftp_get error" begin
-#             options = RequestOptions(ssl=false, active_mode=false, username=user, passwd=pswd, hostname=host)
-#             @test_throws FTPClientError ftp_get(file_name, options)
-#         end
 
-#         @testset "ftp_command error" begin
-#             options = RequestOptions(ssl=false, active_mode=false, username=user, passwd=pswd, hostname=host)
-#             @test_throws FTPClientError ftp_command("NLST", options)
-#         end
+# ftp_put error
+# options = RequestOptions(ssl=false, active_mode=false, username=user, passwd=pswd, hostname=host)
+# open(upload_file_name) do file
+# @test_throws FTPClientError ftp_put(upload_file_name, file, options)
 
-#     end
+# ftp_get error
+# @test_throws FTPClientError ftp_get(file_name, options)
 
-#     @testset "Testing for client failure in FTPObject.jl" begin
+# ftp_command error
+# @test_throws FTPClientError ftp_command("NLST", options)
 
-#         @testset "FTP object error when connecting" begin
-             @test_throws FTPClientError FTP(ssl=false, user=user, pswd=pswd, host="not a host")
-#         end
 
-#         @testset "FTP object error when downloading, blocking" begin
-#             ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
-#             @test_throws FTPClientError download(ftp, file_name)
-#             close(ftp)
-#         end
+# FTP object error when downloading, blocking
+# ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
+# @test_throws FTPClientError download(ftp, file_name)
+# close(ftp)
 
-#         @testset "FTP object error when uploading, blocking" begin
-#             ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
-#             @test_throws FTPClientError upload(ftp, upload_file_name)
-#             close(ftp)
-#         end
+# FTP object error when uploading, blocking
+# ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
+# @test_throws FTPClientError upload(ftp, upload_file_name)
+# close(ftp)
 
-#         @testset "FTP object error when changing working directory" begin
-             ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
-             @test_throws FTPClientError cd(ftp, "testdir")
-             Base.close(ftp)
-#         end
+# FTP object error when changing working directory
+# ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
+# @test_throws FTPClientError cd(ftp, "test_directory")
+# Base.close(ftp)
 
-#         @testset "FTP object error when getting path of working directory" begin
-#             ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
-#             @test_throws FTPClientError pwd(ftp)
-#             close(ftp)
-#         end
+# FTP object error when getting path of working directory
+# ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
+# @test_throws FTPClientError pwd(ftp)
+# close(ftp)
 
-#         @testset "FTP object error when removing a file that doesn't exist" begin
-             ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
-             @test_throws FTPClientError rm(ftp, "not_a_file")
-             Base.close(ftp)
-#         end
+# FTP object error when removing a file that doesn't exist
+ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
+@test_throws FTPClientError rm(ftp, "not_a_file")
+Base.close(ftp)
 
-#         @testset "FTP object error when removing a directory that doesn't exist" begin
-             ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
-             @test_throws FTPClientError rmdir(ftp, "not_a_directory")
-             Base.close(ftp)
-#         end
 
-#         @testset "FTP object error when creating a bad directory" begin
-             ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
-             @test_throws FTPClientError mkdir(ftp, "")
-             Base.close(ftp)
-#         end
+# FTP object error when removing a directory that doesn't exist
+ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
+@test_throws FTPClientError rmdir(ftp, "not_a_directory")
+Base.close(ftp)
 
-#         @testset "FTP object error when moving bad directories" begin
-             ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
-             @test_throws FTPClientError mv(ftp, "", "")
-             Base.close(ftp)
-#         end
+# FTP object error when creating a bad directory
+ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
+@test_throws FTPClientError mkdir(ftp, "")
+Base.close(ftp)
 
-#         @testset "FTP object error when moving bad directories 2" begin
-             ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
-             @test_throws FTPClientError mv(ftp, file_name, "")
-             Base.close(ftp)
-#         end
 
-#         @testset "FTP object error with LIST command" begin
-#             undo_errors()
-#             ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
-#             set_list_error()
-#             @test_throws FTPClientError readdir(ftp)
-#             close(ftp)
-#         end
+# FTP object error when moving bad directories
+ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
+@test_throws FTPClientError mv(ftp, "", "")
+Base.close(ftp)
 
-#     end
 
-#     ftp_cleanup()
+# FTP object error when moving bad directories 2
+ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
+@test_throws FTPClientError mv(ftp, file_name, "")
+Base.close(ftp)
 
-# end
+# FTP object error with LIST command
+# undo_errors()
+# ftp = FTP(ssl=false, user=user, pswd=pswd, host=host)
+# set_list_error()
+# @test_throws FTPClientError readdir(ftp)
+# close(ftp)
+
+ftp_cleanup()
+
+close(server)
