+++ test/runtests.jl
@@ -1,200 +1,37 @@
-push!(LOAD_PATH, "./src")
-
+import Compat: readstring
 using FTPClient
-if VERSION >= v"0.5-"
-    using Base.Test
-else
-    using BaseTestNext
-    const Test = BaseTestNext
-    typealias Future RemoteRef
-end
-using JavaCall
-using LibCURL
-using Compat
+using Base.Test
 
-function start_server()
-    port = jcall(MockFTPServerJulia, "setUp", jint, ())
-    @assert port > 0 "start_server failed"
-    port
-end
 
-function stop_server()
-    result = jcall(MockFTPServerJulia, "tearDown", jboolean, ())
-    @assert result == 1 "stop_server failed"
-end
+include("server/server.jl")
+include("utils.jl")
 
-function set_user(name::AbstractString, password::AbstractString, home_dir::AbstractString)
-    result = jcall(MockFTPServerJulia, "setUser", jboolean, (JString, JString, JString), name, password, home_dir)
-    @assert result == 1 "set_user failed"
-end
+setup_server()
+ftp_init()
+server = FTPServer()
 
-function set_file(name::AbstractString, content::AbstractString)
-    result = jcall(MockFTPServerJulia, "setFile", jboolean, (JString, JString), name, content)
-    @assert result == 1 "set_file failed"
-end
 
-function set_byte_file(name::AbstractString, content::AbstractString)
-    result = jcall(MockFTPServerJulia, "setByteFile", jboolean, (JString, JString), name, content)
-    @assert result == 1 "set_byte_file failed"
-end
+user = "user"
+host = hostname(server)
+pswd = "passwd"
 
-function set_directory(name::AbstractString)
-    result = jcall(MockFTPServerJulia, "setDirectory", jboolean, (JString,), name)
-    @assert result == 1 "set_directory failed"
-end
+testdir = "test_dir"
+upload_file = "test_upload.txt"
+download_file = "test_download.txt"
+tempfile(upload_file)
+tempfile(joinpath(ROOT,download_file))
 
-function set_command_response(request::AbstractString, code::Integer, reponse::AbstractString)
-    result = jcall(MockFTPServerJulia, "setCommandResponse", jboolean, (JString, jint, JString,), request, code, reponse)
-    @assert result == 1 "set_command_response failed"
-end
+cleanup_file(download_file)
+cleanup_file(joinpath(ROOT, upload_file))
 
-function set_errors()
-    result = jcall(MockFTPServerJulia, "setErrors", jboolean, ())
-    @assert result == 1 "set_errors failed"
-end
+include("ssl.jl")
+include("non_ssl.jl")
+include("ftp_object.jl")
 
-function set_list_error()
-    result = jcall(MockFTPServerJulia, "setListError", jboolean, ())
-    @assert result == 1 "set_list_error failed"
-end
+cleanup_file(upload_file)
+cleanup_file(joinpath(ROOT, download_file))
 
-function set_type_error()
-    result = jcall(MockFTPServerJulia, "setTypeError", jboolean, ())
-    @assert result == 1 "set_type_error failed"
-end
+ftp_cleanup()
+teardown_server()
 
-function undo_errors()
-    result = jcall(MockFTPServerJulia, "undoErrors", jboolean, ())
-    @assert result == 1 "undo_errors failed"
-end
-
-function remove(path::AbstractString)
-    result = jcall(MockFTPServerJulia, "remove", jboolean, (JString,), path)
-    @assert result == 1 "remove failed"
-end
-
-function file_exists(path::AbstractString)
-    return jcall(MockFTPServerJulia, "fileExists", jboolean, (JString,), path) == 1
-end
-
-function directory_exists(path::AbstractString)
-    return jcall(MockFTPServerJulia, "directoryExists", jboolean, (JString,), path) == 1
-end
-
-function get_file_contents(path::AbstractString)
-    return bytestring(jcall(MockFTPServerJulia, "getFileContents", JString, (JString,), path))
-end
-
-function get_byte_file_contents(path::AbstractString)
-    return bytestring(jcall(MockFTPServerJulia, "getByteFileContents", JString, (JString,), path))
-end
-
-function set_files()
-    set_file("/" * file_name, file_contents)
-    set_file("/" * directory_name * "/" * file_name2, file_contents)
-    set_byte_file("/" * byte_file_name, byte_file_contents)
-
-    @assert file_exists("/" * file_name)
-    @assert directory_exists("/" * directory_name)
-    @assert file_exists("/" * directory_name * "/" * file_name2)
-    @assert file_exists("/" * byte_file_name)
-
-    # Making it's not always true
-    @assert !directory_exists("this directory does not exist")
-    @assert !file_exists("this file does not exists")
-    @assert !directory_exists("/" * file_name)
-    @assert !file_exists("/" * directory_name)
-end
-
-new_file = "new_name.txt"
-testdir = "testdir"
-host = "localhost"
-original_host = host
-user = "test"
-pswd = "test"
-home_dir = "/"
-file_name = "test_download.txt"
-file_name2 = "test_download2.txt"
-directory_name = "test_directory"
-file_size = rand(1:100)
-file_contents = randstring(file_size)
-@unix_only byte_file_contents = string("466F6F426172", "0D0A", "466F6F426172")
-@unix_only byte_file_contents_ascii_transfer = string("466F6F426172", "0A", "466F6F426172")
-@windows_only byte_file_contents = string("466F6F426172", "0A", "466F6F426172", "1A1A1A")
-byte_file_name = randstring(20)
-byte_upload_file_name = randstring(21)
-upload_file_name = "test_upload.txt"
-upload_file_contents = randstring(100)
-open(upload_file_name, "w") do f
-    write(f, upload_file_contents)
-end
-
-@unix_only upload_local_byte_file_contents = string("466F6F426172", "0A", "466F6F426172")
-@unix_only upload_local_byte_file_contents_ascii_transfer = string("466F6F426172", "0D0A", "466F6F426172")
-@windows_only upload_local_byte_file_contents = string("466F6F426172", "0D0A", "466F6F426172")
-@windows_only upload_local_byte_file_contents_ascii_transfer = string("466F6F426172", "0A", "466F6F426172")
-
-if (length(ARGS) >= 1 && ARGS[1] == "true")
-    test_ssl = true
-else
-    test_ssl = false
-end
-
-if (length(ARGS) >= 2 && ARGS[2] == "true")
-    test_implicit = true
-else
-    test_implicit = false
-end
-
-if (length(ARGS) == 4)
-    user = ARGS[3]
-    pswd = ARGS[4]
-end
-
-if (test_ssl)
-    if (test_implicit)
-        fp = joinpath(dirname(@__FILE__), "test_implicit_ssl.jl")
-        println("$fp ...\n")
-        include(fp)
-    else
-        fp = joinpath(dirname(@__FILE__), "test_explicit_ssl.jl")
-        println("$fp ...\n")
-        include(fp)
-    end
-else
-    # Start Java, and point to the class in this directory
-    pkg_dir = joinpath(dirname(@__FILE__), "..")
-    JavaCall.init([
-        "-Djava.class.path=$(joinpath(pkg_dir, "test"))",
-        "-Djava.ext.dirs=$(joinpath(pkg_dir, "deps", "ext"))",
-    ])
-    MockFTPServerJulia = @jimport MockFTPServerJulia
-
-    set_user(user, pswd, home_dir)
-    set_files()
-    set_command_response("AUTH", 230, "Login successful.")
-    port = start_server()
-    host = "$original_host:$port"
-
-    # Note: If LibCURL complains that the server doesn't listen it probably means that
-    # the MockFtpServer isn't ready to accept connections yet.
-
-    @testset "FTPCLient" begin
-
-        include("test_non_ssl.jl")
-        include("test_ftp_object.jl")
-
-        # Basic commands will now error
-        set_errors()
-
-        include("test_client_errors.jl")
-
-    end
-
-    stop_server()
-    JavaCall.destroy()
-
-end
-
-# Done testing
-rm(upload_file_name)
+close(server)
