+++ test/utils.jl
@@ -16,3 +16,17 @@ function is_header_equal(original::AbstractString, expected::AbstractString)
         return original == expected
     end
 end
+
+function cleanup_file(filename::AbstractString)
+    if (isfile(filename))
+        rm(filename)
+    end
+    @test !isfile(filename)
+end
+
+function cleanup_dir(dirname::AbstractString, recursive = true)
+    if (isdir(dirname))
+        rm(dirname)
+    end
+    @test !isdir(dirname)
+end
